+++ database/migrations/2023_06_22_053739_create_blood_types_table.php
@@ -9,7 +9,7 @@ return new class extends Migration
   public function up(): void {
     Schema::create('blood_types', function (Blueprint $table) {
       $table->id();
-      $table->integer('user_id')->unsinged();
+      $table->bigInteger('user_id')->unsinged();
       $table->string('name');
       $table->timestamps();
     });
