+++ database/migrations/2023_06_22_055851_create_experiences_table.php
@@ -9,15 +9,15 @@ return new class extends Migration
   public function up(): void {
     Schema::create('experiences', function (Blueprint $table) {
       $table->id();
-      $table->integer('person_id')->unsigned();
+      $table->bigInteger('person_id')->unsigned();
       $table->string('position');
       $table->string('company');
       $table->date('from');
       $table->date('to')->nullable();
       $table->decimal('salary', 8, 2)->nullable();
-      $table->integer('salary_type_id')->unsinged();
+      $table->bigInteger('salary_type_id')->unsinged();
       $table->tinyInteger('sg')->unsgined()->nullable();
-      $table->tinyInteger('work_status_id')->unsgined();
+      $table->bigInteger('work_status_id')->unsgined();
         // [1] hourly, [2] daily, [3] monthly
       $table->boolean('is_government')->default(0);
       $table->timestamps();
