+++ database/migrations/2023_06_22_080958_create_jobs_table.php
@@ -9,14 +9,14 @@ return new class extends Migration
   public function up(): void {
     Schema::create('jobs', function (Blueprint $table) {
       $table->id();
-      $table->integer('user_id')->unsinged();
+      $table->bigInteger('user_id')->unsinged();
       $table->string('title');
       $table->date('posted_at');
       $table->date('unposted_at');
-      $table->integer('job_location_id')->unsigned();
-      $table->integer('job_department_id')->unsigned();
-      $table->integer('job_type_id')->unsigned();
-      $table->integer('job_status_id')->unsigned();
+      $table->bigInteger('job_location_id')->unsigned();
+      $table->bigInteger('job_department_id')->unsigned();
+      $table->bigInteger('job_type_id')->unsigned();
+      $table->bigInteger('job_status_id')->unsigned();
       $table->string('requirements');
       $table->string('responsibilities');
       $table->string('qualifications');
